Time: 0 ms (100%), Space: 21.9 MB (39.48%) - LeetHub

+++ 0268-missing-number/0268-missing-number.cpp
@@ -1,8 +1,8 @@
 class Solution {
 public:
     int missingNumber(vector<int>& nums) {
-        int sum1 = 0;
-        int sum2 = 0;
+        short int sum1 = 0;
+        short int sum2 = 0;
 
         for (int i = 1; i <= nums.size(); i++)
         {
